Snake Game

Plays the classic game Snake. A full walkthrough of the creation of this ROM can be found in vignette("snake").

Usage

Snake

Format

A game ROM; see vignette("engine").

How to Play

WASD to move the snake. ENTER must be pressed after each input for it to be registered.

Follow these steps:

quickload(Snake)

Follow console prompt.

Open new R session and load rcade with library(rcade).

inputs.listen() in that new session.

Use WASD to control snake.

When you run quickload(), the game will ask if you want to keep the default settings or input your own. The default settings are:

5 frames per second
32x16 bounding area
10 starting snake length