Test an Animated Sprite

Plays a sprite's animation into the console. ^C to stop.

Usage

render.test_animation(spritename, sprites)

Arguments

spritename

Name of sprite in sprites

sprites

List of sprites, e.g. ROM$sprites.

Details

This function is intended to aid in the sprite creation process.

It handles sprite$next_animation properly, so it can be used to test multi-stage animations too.

todo: examples with mario runcycle